+++ src/pharmpy/modeling/__init__.py
@@ -1,5 +1,6 @@
 from pharmpy.modeling.block_rvs import create_joint_distribution, split_joint_distribution
 from pharmpy.modeling.common import (
+        print_model_code,
     add_covariance_step,
     add_estimation_step,
     append_estimation_step_options,
@@ -198,6 +199,7 @@ __all__ = [
     'predict_influential_individuals',
     'predict_influential_outliers',
     'predict_outliers',
+    'print_model_code',
     'print_model_symbols',
     'read_model',
     'read_model_from_database',